Refactoring based on review feedback
